requirements.txt
requests>=2.31.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
httpx>=0.25.0
python-dateutil>=2.8.0

playwright>=1.50.0
playwright-stealth>=1.0.6
sentence-transformers>=3.0.0
flask>=3.0.0
